Split RabbitMQPlugin into ExtractorRoutingService and MessageService (#155)
* Initial split into services

* single MessageService file

* Add stubbed EventSinkService (using fake MessageService)

* Included an example of how to call EventSinkService from within an HTML template

* Wiring updates

* wiring for ExtractorRoutingService

* more wiring

* Sync to switch branches

* further wiring of new services

* more wiring

* enough updates to compile!

* add Singleton marker

* initialize factory to account for lack of onStart

* Filled out EventSinkService impl

* Adjust direct calls to Amplitude to use EventSinkService

* add noop for MessageService

* fix: cleanup commented out yucky code

* Added missing tracking for file uploads

* Remove direct usage of Amplitude Javascript SDK

* Fix logFileUploadEvent, starting to think about auth question

* add manual submission event

* Update RabbitMQMessageService.scala

* Fix failing test case / wrong package name

* Fix failing test case, revert package name change

* Parameterize exchange_type (default value is "topic")

* Removed old call to Amplitude. It was breaking the upload page. File div would disappear after upload even though the file was successfully added.

* Added rabbitmq info back to status. Created a getInfo() function to let the implementation specify what to use.
Might want to use the same pattern for future plugin -> services conversions.
