+++ app/controllers/Files.scala
@@ -51,7 +51,9 @@ class Files @Inject() (
   contextLDService: ContextLDService,
   spaces: SpaceService,
   folders: FolderService,
-  appConfig: AppConfigurationService) extends SecuredController {
+  routing: ExtractorRoutingService,
+  appConfig: AppConfigurationService,
+  sinkService: EventSinkService) extends SecuredController {
 
   lazy val chunksize = play.Play.application().configuration().getInt("clowder.chunksize", 1024*1024)
 
@@ -276,6 +278,7 @@ class Files @Inject() (
             // Increment view count for file
             val (view_count, view_date) = files.incrementViews(id, user)
 
+            sinkService.logFileViewEvent(file, user)
             //passing None as the last parameter (list of output formats)
             Future(Ok(views.html.file(file, id.stringify, commentsByFile, previewsWithPreviewer, sectionsWithPreviews,
               extractorsActive, decodedDatasetsContaining.toList, foldersContainingFile,
@@ -464,10 +467,7 @@ class Files @Inject() (
                   }
                 }
                 // submit for extraction
-                current.plugin[RabbitmqPlugin].foreach {
-                  // FIXME dataset not available?
-                  _.fileCreated(f, None, Utils.baseUrl(request), request.apiKey)
-                }
+                routing.fileCreated(f, None, Utils.baseUrl(request).toString, request.apiKey)
 
                 /** *** Inserting DTS Requests   **/
                 val clientIP = request.remoteAddress
@@ -607,10 +607,7 @@ class Files @Inject() (
               val extra = Map("filename" -> f.filename)
               dtsrequests.insertRequest(serverIP, clientIP, f.filename, f.id, fileType, f.length, f.uploadDate)
               /****************************/
-	            current.plugin[RabbitmqPlugin].foreach{
-                // FIXME dataset not available?
-                _.fileCreated(f, None, Utils.baseUrl(request), request.apiKey)
-              }
+              routing.fileCreated(f, None, Utils.baseUrl(request).toString, request.apiKey)
 	            
 	            //for metadata files
 	            if(fileType.equals("application/xml") || fileType.equals("text/xml")){
@@ -712,6 +709,7 @@ class Files @Inject() (
             files.getBytes(id) match {
               case Some((inputStream, filename, contentType, contentLength)) => {
                 files.incrementDownloads(id, user)
+                sinkService.logFileDownloadEvent(file, user)
 
                 request.headers.get(RANGE) match {
                   case Some(value) => {
@@ -987,9 +985,7 @@ class Files @Inject() (
               _.dump(DumpOfFile(uploadedFile.ref.file, f.id.toString, nameOfFile))
             }
 
-            current.plugin[RabbitmqPlugin].foreach {
-              _.multimediaQuery(f.id, f.contentType, f.length.toString, Utils.baseUrl(request), request.apiKey)
-            }
+            routing.multimediaQuery(f.id, f.contentType, f.length.toString, Utils.baseUrl(request), request.apiKey)
 
             //for metadata files
             if (fileType.equals("application/xml") || fileType.equals("text/xml")) {
@@ -1084,9 +1080,7 @@ class Files @Inject() (
             val id = f.id
             val extra = Map("filename" -> f.filename, "action" -> "upload")
 
-            current.plugin[RabbitmqPlugin].foreach {
-              _.fileCreated(f, host, request.apiKey)
-            }
+            routing.fileCreated(f, host, request.apiKey)
 
             //for metadata files
             if (fileType.equals("application/xml") || fileType.equals("text/xml")) {
@@ -1222,10 +1216,8 @@ class Files @Inject() (
                     }
 
                     // notify extractors that a file has been uploaded and added to a dataset
-                    current.plugin[RabbitmqPlugin].foreach { rabbitMQ =>
-                      rabbitMQ.fileCreated(f, Some(dataset), Utils.baseUrl(request), request.apiKey)
-                      rabbitMQ.fileAddedToDataset(f, dataset, Utils.baseUrl(request), request.apiKey)
-                    }
+                    routing.fileCreated(f, Some(dataset), Utils.baseUrl(request).toString, request.apiKey)
+                    routing.fileAddedToDataset(f, dataset, Utils.baseUrl(request), request.apiKey)
 
                     // add file to RDF triple store if triple store is used
                     if (fileType.equals("application/xml") || fileType.equals("text/xml")) {
