+++ app/services/DI.scala
@@ -56,6 +56,7 @@ class ConfigurationModule extends AbstractModule {
     bind(classOf[SelectionService]).to(get("service.select", "services.mongodb.MongoDBSelectionService"))
     bind(classOf[FileLinkService]).to(get("service.FileLinkService", "services.mongodb.MongoDBFileLinkService"))
     bind(classOf[LogService]).to(get("service.LogService", "services.GraylogService"))
+    bind(classOf[MessageService]).to(get("service.MessageService", "services.RabbitMQMessageService"))
   }
 
   protected def get[T](key: String, missing: String) : Class[T] = {
