+++ app/views/datasets/uploadFiles.scala.html
@@ -127,7 +127,6 @@ of the blueimp jQuery File Upload library. An open source project located here: 
 <!-- The template to display files available for download -->
 <script id="template-download" type="text/x-tmpl">
     {% for (var i=0, file; file=o.files[i]; i++) { %}
-        {% logEvent(file.size); %}
         <tr class="template-download fade">
             {% if (file.deleteUrl) { %}
             <td>
@@ -174,21 +173,6 @@ of the blueimp jQuery File Upload library. An open source project located here: 
 <!-- The Templates plugin is included to render the upload/download listings - downloaded to make the resource local -->
 <script src="@routes.Assets.at("javascripts/file-uploader/tmpl.min.js")"></script>
 
-<script>
-    // Inject some info into templates for tracking
-    @if(AppConfiguration.getAmplitudeApiKey != "") {
-        tmpl.helper += ',logEvent=function(size){' +
-         'amplitude.logEvent("upload", {' +
-            '"dataset_id":"@dataset.id",' +
-            '"dataset_name":"@dataset.name",' +
-            '"uploader_id":"@user.get.id",' +
-            '"uploader_name":"@user.get.getMiniUser.fullName",' +
-            '"length":size,' +
-            '});}'
-    } else {
-            tmpl.helper += ',logEvent=function(){}'
-    }
-</script>
 <!-- The Load Image plugin is included for the preview images and image resizing functionality - downloaded to make the resource local -->
 <script src="@routes.Assets.at("javascripts/file-uploader/load-image.all.min.js")"></script>
 <!-- The Canvas to Blob plugin is included for image resizing functionality - downloaded to make the resource local -->
