+++ app/views/file.scala.html
@@ -72,20 +72,6 @@
 	<script src="@routes.Assets.at("javascripts/deleteUtils.js")" type="text/javascript"></script>
 	<script src="@routes.Assets.at("javascripts/sectionsListProcess.js")" type="text/javascript"></script>
 	<script src="@routes.Assets.at("javascripts/files/datasets.js")" type="text/javascript"></script>
-	@if(AppConfiguration.getAmplitudeApiKey != "") {
-		<script>
-				amplitude.getInstance().logEvent("view_resource", {
-					"type": "file",
-					"resource_id": '@file.id',
-					"resource_name": '@file.filename',
-					"author_id": '@file.author.id',
-					"author_name": '@file.author.fullName',
-					"viewer_id": '@user.get.id',
-					"viewer_name": '@user.get.getMiniUser.fullName',
-					"length": @file.length
-				});
-		</script>
-	}
 
 	<script>
 			var spaceId = "@spaceId";
@@ -402,8 +388,7 @@
 							objectType="file"><span class="glyphicon glyphicon-star-empty"></span> Unfollow
 							</button>
 						}
-                        @if(Permission.checkPermission(Permission.EditFile, ResourceRef(ResourceRef.file, file.id))
-                                && play.api.Play.current.plugin[services.RabbitmqPlugin].isDefined) {
+                        @if(Permission.checkPermission(Permission.EditFile, ResourceRef(ResourceRef.file, file.id))) {
                             <a href="@routes.Extractors.submitFileExtraction(UUID(id))" class="btn btn-link" >
                                 <span class="glyphicon glyphicon-send"></span>
                                 Submit for extraction
